Add location to AST
I'm working on a tool that allows people to analyze regular expressions,
and as part of that I'd like to be able to output error messages with
character positions, and highlight relevant portions of the regular
expression.

I chose to add a `loc` property to each node to encapsulate the change,
but am open to other approaches if you'd prefer.